Revert "Feature/130/mouse movement interface"
