+++ docs/classes/bresenham.html
@@ -95,7 +95,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/util/bresenham.class.ts#L4">util/bresenham.class.ts:4</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/util/bresenham.class.ts#L4">util/bresenham.class.ts:4</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
