+++ docs/classes/clipboardaction.html
@@ -110,7 +110,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboardy-clipboard-action.class.ts#L4">provider/native/clipboardy-clipboard-action.class.ts:4</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboardy-clipboard-action.class.ts#L4">provider/native/clipboardy-clipboard-action.class.ts:4</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <a href="clipboardaction.html" class="tsd-signature-type">ClipboardAction</a></h4>
@@ -131,7 +131,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/clipboardactionprovider.html">ClipboardActionProvider</a>.<a href="../interfaces/clipboardactionprovider.html#clear">clear</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboardy-clipboard-action.class.ts#L19">provider/native/clipboardy-clipboard-action.class.ts:19</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboardy-clipboard-action.class.ts#L19">provider/native/clipboardy-clipboard-action.class.ts:19</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">&gt;</span></h4>
@@ -149,7 +149,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/clipboardactionprovider.html">ClipboardActionProvider</a>.<a href="../interfaces/clipboardactionprovider.html#copy">copy</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboardy-clipboard-action.class.ts#L23">provider/native/clipboardy-clipboard-action.class.ts:23</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboardy-clipboard-action.class.ts#L23">provider/native/clipboardy-clipboard-action.class.ts:23</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -173,7 +173,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/clipboardactionprovider.html">ClipboardActionProvider</a>.<a href="../interfaces/clipboardactionprovider.html#hastext">hasText</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboardy-clipboard-action.class.ts#L8">provider/native/clipboardy-clipboard-action.class.ts:8</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboardy-clipboard-action.class.ts#L8">provider/native/clipboardy-clipboard-action.class.ts:8</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">&gt;</span></h4>
@@ -191,7 +191,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/clipboardactionprovider.html">ClipboardActionProvider</a>.<a href="../interfaces/clipboardactionprovider.html#paste">paste</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboardy-clipboard-action.class.ts#L34">provider/native/clipboardy-clipboard-action.class.ts:34</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboardy-clipboard-action.class.ts#L34">provider/native/clipboardy-clipboard-action.class.ts:34</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">&gt;</span></h4>
