+++ docs/classes/imagereader.html
@@ -102,7 +102,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/datasource.html">DataSource</a>.<a href="../interfaces/datasource.html#load">load</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/opencv/image-reader.class.ts#L6">provider/opencv/image-reader.class.ts:6</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/opencv/image-reader.class.ts#L6">provider/opencv/image-reader.class.ts:6</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
