+++ docs/classes/keyboard.html
@@ -116,7 +116,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L24">keyboard.class.ts:24</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L24">keyboard.class.ts:24</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -152,7 +152,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L73">keyboard.class.ts:73</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L73">keyboard.class.ts:73</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -190,7 +190,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L95">keyboard.class.ts:95</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L95">keyboard.class.ts:95</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -228,7 +228,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L44">keyboard.class.ts:44</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L44">keyboard.class.ts:44</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -264,7 +264,7 @@
 					<div class="tsd-signature tsd-kind-icon">config<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">object</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L19">keyboard.class.ts:19</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L19">keyboard.class.ts:19</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -278,7 +278,7 @@
 						<div class="tsd-signature tsd-kind-icon">auto<wbr>Delay<wbr>Ms<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> = 300</span></div>
 						<aside class="tsd-sources">
 							<ul>
-								<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/keyboard.class.ts#L23">keyboard.class.ts:23</a></li>
+								<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/keyboard.class.ts#L23">keyboard.class.ts:23</a></li>
 							</ul>
 						</aside>
 						<div class="tsd-comment tsd-typography">
