+++ docs/classes/matchresult.html
@@ -107,7 +107,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/match-result.class.ts#L3">match-result.class.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/match-result.class.ts#L3">match-result.class.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -132,7 +132,7 @@
 					<div class="tsd-signature tsd-kind-icon">confidence<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/match-result.class.ts#L5">match-result.class.ts:5</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/match-result.class.ts#L5">match-result.class.ts:5</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -142,7 +142,7 @@
 					<div class="tsd-signature tsd-kind-icon">location<span class="tsd-signature-symbol">:</span> <a href="region.html" class="tsd-signature-type">Region</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/match-result.class.ts#L6">match-result.class.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/match-result.class.ts#L6">match-result.class.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
