+++ docs/classes/scaledmatchresult.html
@@ -109,7 +109,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides <a href="matchresult.html">MatchResult</a>.<a href="matchresult.html#constructor">constructor</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/scaled-match-result.class.ts#L4">scaled-match-result.class.ts:4</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/scaled-match-result.class.ts#L4">scaled-match-result.class.ts:4</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -138,7 +138,7 @@
 					<aside class="tsd-sources">
 						<p>Overrides <a href="matchresult.html">MatchResult</a>.<a href="matchresult.html#confidence">confidence</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/scaled-match-result.class.ts#L5">scaled-match-result.class.ts:5</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/scaled-match-result.class.ts#L5">scaled-match-result.class.ts:5</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -149,7 +149,7 @@
 					<aside class="tsd-sources">
 						<p>Overrides <a href="matchresult.html">MatchResult</a>.<a href="matchresult.html#location">location</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/scaled-match-result.class.ts#L7">scaled-match-result.class.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/scaled-match-result.class.ts#L7">scaled-match-result.class.ts:7</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -159,7 +159,7 @@
 					<div class="tsd-signature tsd-kind-icon">scale<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/scaled-match-result.class.ts#L6">scaled-match-result.class.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/scaled-match-result.class.ts#L6">scaled-match-result.class.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
