+++ docs/classes/screenaction.html
@@ -112,7 +112,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L16">provider/native/libnut-screen-action.class.ts:16</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L16">provider/native/libnut-screen-action.class.ts:16</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <a href="screenaction.html" class="tsd-signature-type">ScreenAction</a></h4>
@@ -133,7 +133,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#grabscreen">grabScreen</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L21">provider/native/libnut-screen-action.class.ts:21</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L21">provider/native/libnut-screen-action.class.ts:21</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><a href="image.html" class="tsd-signature-type">Image</a><span class="tsd-signature-symbol">&gt;</span></h4>
@@ -151,7 +151,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#grabscreenregion">grabScreenRegion</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L45">provider/native/libnut-screen-action.class.ts:45</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L45">provider/native/libnut-screen-action.class.ts:45</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -175,7 +175,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#highlightscreenregion">highlightScreenRegion</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L70">provider/native/libnut-screen-action.class.ts:70</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L70">provider/native/libnut-screen-action.class.ts:70</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -205,7 +205,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#screenheight">screenHeight</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L89">provider/native/libnut-screen-action.class.ts:89</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L89">provider/native/libnut-screen-action.class.ts:89</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">&gt;</span></h4>
@@ -223,7 +223,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#screensize">screenSize</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L100">provider/native/libnut-screen-action.class.ts:100</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L100">provider/native/libnut-screen-action.class.ts:100</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><a href="region.html" class="tsd-signature-type">Region</a><span class="tsd-signature-symbol">&gt;</span></h4>
@@ -241,7 +241,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/screenactionprovider.html">ScreenActionProvider</a>.<a href="../interfaces/screenactionprovider.html#screenwidth">screenWidth</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/libnut-screen-action.class.ts#L78">provider/native/libnut-screen-action.class.ts:78</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/libnut-screen-action.class.ts#L78">provider/native/libnut-screen-action.class.ts:78</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">&gt;</span></h4>
