+++ docs/classes/templatematchingfinder.html
@@ -108,7 +108,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/opencv/template-matching-finder.class.ts#L65">provider/opencv/template-matching-finder.class.ts:65</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/opencv/template-matching-finder.class.ts#L65">provider/opencv/template-matching-finder.class.ts:65</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -134,7 +134,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/opencv/template-matching-finder.class.ts#L175">provider/opencv/template-matching-finder.class.ts:175</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/opencv/template-matching-finder.class.ts#L175">provider/opencv/template-matching-finder.class.ts:175</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -160,7 +160,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/opencv/template-matching-finder.class.ts#L72">provider/opencv/template-matching-finder.class.ts:72</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/opencv/template-matching-finder.class.ts#L72">provider/opencv/template-matching-finder.class.ts:72</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
