+++ docs/enums/filetype.html
@@ -91,7 +91,7 @@
 					<div class="tsd-signature tsd-kind-icon">JPG<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;.jpg&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/file-type.enum.ts#L6">file-type.enum.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/file-type.enum.ts#L6">file-type.enum.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -101,7 +101,7 @@
 					<div class="tsd-signature tsd-kind-icon">PNG<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;.png&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/file-type.enum.ts#L5">file-type.enum.ts:5</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/file-type.enum.ts#L5">file-type.enum.ts:5</a></li>
 						</ul>
 					</aside>
 				</section>
