+++ docs/index.html
@@ -270,7 +270,6 @@ describe(<span class="hljs-string">"Basic test"</span>, () =&gt; {
 					<li><input checked="" disabled="" type="checkbox"> findOnScreen</li>
 					<li><input checked="" disabled="" type="checkbox"> waitFor</li>
 					<li><input checked="" disabled="" type="checkbox"> Hooks to trigger actions based on images</li>
-					<li><input checked="" disabled="" type="checkbox"> Highlighting screen regions</li>
 				</ul>
 				<a href="#integration" id="integration" style="color: inherit; text-decoration: none;">
 					<h2>Integration</h2>
@@ -387,9 +386,6 @@ describe(<span class="hljs-string">"Basic test"</span>, () =&gt; {
 						<a href="interfaces/datasource.html" class="tsd-kind-icon">Data<wbr>Source</a>
 					</li>
 					<li class=" tsd-kind-interface">
-						<a href="interfaces/easingfunction.html" class="tsd-kind-icon">Easing<wbr>Function</a>
-					</li>
-					<li class=" tsd-kind-interface">
 						<a href="interfaces/finderinterface.html" class="tsd-kind-icon">Finder<wbr>Interface</a>
 					</li>
 					<li class=" tsd-kind-interface">
@@ -415,15 +411,6 @@ describe(<span class="hljs-string">"Basic test"</span>, () =&gt; {
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="globals.html#busywaitfornanoseconds" class="tsd-kind-icon">busy<wbr>Wait<wbr>For<wbr>Nano<wbr>Seconds</a>
-					</li>
-					<li class=" tsd-kind-function">
-						<a href="globals.html#calculatebasestepduration" class="tsd-kind-icon">calculate<wbr>Base<wbr>Step<wbr>Duration</a>
-					</li>
-					<li class=" tsd-kind-function">
-						<a href="globals.html#calculatemovementtimesteps" class="tsd-kind-icon">calculate<wbr>Movement<wbr>Timesteps</a>
-					</li>
-					<li class=" tsd-kind-function">
-						<a href="globals.html#calculatestepduration" class="tsd-kind-icon">calculate<wbr>Step<wbr>Duration</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="globals.html#centerof" class="tsd-kind-icon">center<wbr>Of</a>
