+++ docs/interfaces/__global.jest.matchers.html
@@ -107,7 +107,7 @@
 					<div class="tsd-signature tsd-kind-icon">to<wbr>BeAt<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">(</span>position<span class="tsd-signature-symbol">: </span><a href="../classes/point.html" class="tsd-signature-type">Point</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-symbol">{}</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/expect/jest.matcher.function.ts#L10">expect/jest.matcher.function.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/expect/jest.matcher.function.ts#L10">expect/jest.matcher.function.ts:10</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-type-declaration">
@@ -140,7 +140,7 @@
 					<div class="tsd-signature tsd-kind-icon">to<wbr>BeIn<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">(</span>region<span class="tsd-signature-symbol">: </span><a href="../classes/region.html" class="tsd-signature-type">Region</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-symbol">{}</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/expect/jest.matcher.function.ts#L11">expect/jest.matcher.function.ts:11</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/expect/jest.matcher.function.ts#L11">expect/jest.matcher.function.ts:11</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-type-declaration">
@@ -173,7 +173,7 @@
 					<div class="tsd-signature tsd-kind-icon">to<wbr>Show<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">(</span>needle<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span>, confidence<span class="tsd-signature-symbol">?: </span><span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-symbol">{}</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/expect/jest.matcher.function.ts#L12">expect/jest.matcher.function.ts:12</a></li>
+							<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/expect/jest.matcher.function.ts#L12">expect/jest.matcher.function.ts:12</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-type-declaration">
