+++ docs/interfaces/clipboardactionprovider.html
@@ -111,7 +111,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboard-action-provider.interface.ts#L17">provider/native/clipboard-action-provider.interface.ts:17</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboard-action-provider.interface.ts#L17">provider/native/clipboard-action-provider.interface.ts:17</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -134,7 +134,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboard-action-provider.interface.ts#L24">provider/native/clipboard-action-provider.interface.ts:24</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboard-action-provider.interface.ts#L24">provider/native/clipboard-action-provider.interface.ts:24</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -165,7 +165,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboard-action-provider.interface.ts#L10">provider/native/clipboard-action-provider.interface.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboard-action-provider.interface.ts#L10">provider/native/clipboard-action-provider.interface.ts:10</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -188,7 +188,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/clipboard-action-provider.interface.ts#L31">provider/native/clipboard-action-provider.interface.ts:31</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/clipboard-action-provider.interface.ts#L31">provider/native/clipboard-action-provider.interface.ts:31</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
