+++ docs/interfaces/datasource.html
@@ -113,7 +113,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/opencv/data-source.interface.ts#L11">provider/opencv/data-source.interface.ts:11</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/opencv/data-source.interface.ts#L11">provider/opencv/data-source.interface.ts:11</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
