+++ docs/interfaces/keyboardactionprovider.html
@@ -112,7 +112,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/keyboard-action-provider.interface.ts#L26">provider/native/keyboard-action-provider.interface.ts:26</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/keyboard-action-provider.interface.ts#L26">provider/native/keyboard-action-provider.interface.ts:26</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -143,7 +143,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/keyboard-action-provider.interface.ts#L33">provider/native/keyboard-action-provider.interface.ts:33</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/keyboard-action-provider.interface.ts#L33">provider/native/keyboard-action-provider.interface.ts:33</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -174,7 +174,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/keyboard-action-provider.interface.ts#L40">provider/native/keyboard-action-provider.interface.ts:40</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/keyboard-action-provider.interface.ts#L40">provider/native/keyboard-action-provider.interface.ts:40</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -205,7 +205,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/keyboard-action-provider.interface.ts#L12">provider/native/keyboard-action-provider.interface.ts:12</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/keyboard-action-provider.interface.ts#L12">provider/native/keyboard-action-provider.interface.ts:12</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -236,7 +236,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/da554bf/lib/provider/native/keyboard-action-provider.interface.ts#L19">provider/native/keyboard-action-provider.interface.ts:19</a></li>
+									<li>Defined in <a href="https://github.com/nut-tree/nut.js/blob/9a70841/lib/provider/native/keyboard-action-provider.interface.ts#L19">provider/native/keyboard-action-provider.interface.ts:19</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
